Updates to Re-enable Standalone Application
The launch file for the standalone application (surf_raster_planner_application.launch) could not be started.  It appears that the name of the application executable had been changed without updating the launch file.  Additionally, there were several files required that were not being installed during the build process.

This commit changes the name in the launch file from noether_node (which I could not find) to surface_raster_planner_application.  In addition, it adds install rules for the launch file and the tool configuration file.