Refactoring: Making Old Behavior the Default
When the ability to rotate the raster direction was added in PR 31, there were two parameters added to the ProcessTool struct and the corresponding ToolPathConfig message type.  The first one was the raster angle.  The second one was a flag to switch between rotating about one of the principal axes and rotating about the mesh's coordinate frame axes.  As indicated in the documentation for the ProcessTool struct, the intent was for this flag to default to true (principal axis).  Unfortunately, ROS messages fill booleans with false (coordinate axis) by default.

This caused previously change code to subtly change without notice.  Some meshes would not be affected, while some would now fail to generate an initial intersection.

This commit refactors the 'raster_wrt_principal_axis' field to be 'raster_wrt_global_axes', as well as adjusting the places that this code is called in the rest of the library.  This will ensure that, as intended, using principal axes is the default in both the struct and the ROS message.